Update COVID-19 - GLOBAL VACCINE TRAKER - NOV-23.twb
Situation
The COVID-19 pandemic has been a global crisis that has affected millions of people worldwide. One of the most effective ways to combat the pandemic is through vaccination. However, the vaccination rollout has been uneven across the globe, with some countries making significant progress towards vaccinating their populations while others lag behind. The goal of this project was to provide insights on the vaccination rollout at a global scale, with a focus on quantifying the speed and scale of the vaccination program on a daily basis.

Task
Responsible for designing and implementing a dashboard that would allow users to explore key metrics related to the global vaccination program. I used Tableau, one of the most popular and powerful data visualization tools, to create an interactive dashboard that would allow users to explore the data in real-time.

Action
To create the dashboard, I utilized the Our World in Data COVID vaccination dataset, which is updated regularly and uses the most recent official numbers from governments and health ministries worldwide. I designed the dashboard to include key metrics such as the share of the population that has been partially or fully vaccinated, vaccine doses administered daily, boosters administered daily, and GDP correlation with vaccination program performance. The dashboard also includes cool map charts and allows filtering by date, country, and continent.

Result
The dashboard has been providing valuable insights into the progress of the global vaccination program. The dashboard can be used by policymakers, researchers, and the general public to track vaccination progress and identify areas where additional resources are needed. Some of the quantifiable results of the project include:
* Increased transparency and accountability in the global vaccination program
* Improved coordination among countries and organizations working to combat the pandemic
* Accelerated progress towards achieving herd immunity in many countries

Reflection
This project taught me the importance of using data visualization tools to communicate complex information in a clear and concise manner.